moving from usize to isize

+++ src/charge_configurations.rs
@@ -7,29 +7,29 @@ use rayon::prelude::*;
 
 
 pub fn closed_charge_configurations_brute_force(
-    n_charge: usize,
-    n_dot: usize,
-    floor_values: ArrayView<usize, Ix1>,
-) -> Array2<usize> {
-    let floor_sum: usize = floor_values.sum();
+    n_charge: isize,
+    n_dot: isize,
+    floor_values: ArrayView<isize, Ix1>,
+) -> Array2<isize> {
+    let floor_sum: isize = floor_values.sum();
 
     if floor_sum > n_charge {
-        return Array2::default((0, n_dot)); // Return an empty array
+        return Array2::default((0, n_dot as usize)); // Return an empty array
     }
 
-    if floor_values.iter().map(|&x| x + 1).sum::<usize>() < n_charge {
-        return Array2::default((0, n_dot)); // Return an empty array
+    if floor_values.iter().map(|&x| x + 1).sum::<isize>() < n_charge {
+        return Array2::default((0, n_dot as usize)); // Return an empty array
     }
 
-    let num_combinations = 2usize.pow(floor_values.len() as u32);
+    let num_combinations = 2isize.pow(floor_values.len() as u32);
     let result = (0..num_combinations)
         .map(|i| {
             let binary_str = format!("{:01$b}", i, floor_values.len());
-            let binary_values: Vec<usize> = binary_str
+            let binary_values: Vec<isize> = binary_str
                 .chars()
-                .map(|c| c.to_digit(2).unwrap() as usize)
+                .map(|c| c.to_digit(2).unwrap() as isize)
                 .collect();
-            let sum: usize = binary_values.iter().sum();
+            let sum: isize = binary_values.iter().sum();
             if sum == n_charge - floor_sum { Some(binary_values) } else { None }
         })
         .filter_map(|x| x)
@@ -40,6 +40,6 @@ pub fn closed_charge_configurations_brute_force(
             comb
         })
         .flatten()
-        .collect::<Vec<usize>>();
-    Array2::from_shape_vec((result.len() / n_dot, n_dot), result).unwrap()
+        .collect::<Vec<isize>>();
+    Array2::from_shape_vec((result.len() / n_dot as usize, n_dot as usize), result).unwrap()
 }
